Delete Quality Mindset.txt
Quality Mindset.
_________________

Quality is a mindset that impacts everyone. It is more than a process or programme to be implemented by workers
 on an assembly line.  It impacts every decision we make, and isn't just limited to the workplace—it impacts us 
at home, work and school.It is the difference between being "good enough" and the best you can be.Quality is an 
age-old concept. Aspiring for the best has been a dominant impulse of humanbeings across civilizations. That is
why we have meticulously built great bath of Indus Valley Civilization or pyramids of Egypt. All the seven monumental 
wonders of the world or hundreds of heritage sites across the world are living examples of human desire for greatest
level of perfection.In modern context, quality has been defined by scholars and practitioners in various ways.Reeves 
and Bednar (1994) define quality as “… value, conformance to specifications,conformance to requirements, fitness for
 use, loss avoidance, and meeting and/or exceeding customers’ expectations.” 

The four pillars required to have a quality mindset
> Vision
> Empathy
> Commitment
> Adaptability

Vision - According to Stephen Covey Vision is the portable version of Covey's "The 7 Effects of Highly Effective People".
This book is full of inspirational and humorous quotes. There's no instructions, no how-tos and no formulas for success.
Instead you will find thought-provoking ideas , workable solutions , insight , and humor from some of the world's most 
influential and inspiring people. Even if you don't want to read the original book, the "Vision" book will encourage you
 to learn valuable lessons of life. 

Empathy - Empathy is the ability to emotionally understand what other people feel, see things from their point of view, and
imagine yourself in their place. Essentially, it is putting yourself in someone else's position and feeling what they must
be feeling.When you see another person suffering, you might be able to instantly envision yourself in the other person's   
place and feel sympathy for what they are going through.While people are generally pretty well-attuned to their own feelings
and emotions, getting into someone else's head can be a bit more difficult. The ability to feel empathy allows people to "walk
a mile in another's shoes," so to speak. It permits people to understand the emotions that others are feeling.For many, seeing
another person in pain and responding with indifference or even outright hostility seems utterly incomprehensible. But the fact
 that some people do respond in such a way clearly demonstrates that empathy is not necessarily a universal response to the
 suffering of others.

Commitment - Making a commitment involves dedicating yourself to something, like a person or a cause. Before you make a commitment,
think carefully. A commitment obligates you to do something.Some commitments are large, like marriage. When you take a job, you're
making a commitment to show up and do the job well, and your employer makes a commitment to pay you. There are smaller commitments 
too. If you said you'd meet a friend at six, that's a commitment — show up or your friend will be mad. You also can speak of commitment
as a quality. Staying after school for a study group shows your commitment to good grades.

Adaptability - Adaptability (Latin: adaptō "fit to, adjust") is a feature of a system or of a process. This word has been put to use 
as a specialised term in different disciplines and in business operations. Word definitions of adaptability as a specialised term differ
little from dictionary definitions. According to Andresen and Gronau[1] adaptability in the field of organizational management can in 
general be seen as an ability to change something or oneself to fit to occurring changes. In ecology, adaptability has been described 
as the ability to cope with unexpected disturbances in the environment.